Enable customized caching in local proxy
For unit testing, it is convenient to be able to control the lifetime
and contents of the cache precisely. To accomplish this, we allow users
to specify a `BRAINTRUST_PROXY_LOCAL_CACHE_PATH` environment variable,
pointing to a JSON file of the cache contents. If specified, the local
nodeJS version of the proxy will use this as the cache instead of redis,
and supports dumping the cache contents back to this file at any point
in time.

Other changes are just cleaning up some function signatures and passing
more hooks through to control caching behavior.