Ensure cache encryption is not reversible (#46)
Fixes https://github.com/braintrustdata/braintrust-proxy/issues/44

As demonstrated in the test script, the previous encryption strategy was
reversible as long as you had the OpenAI key used. At least for my use
case, this violates the guarantees we try to give customers of not
retaining copies of their data.

This switches it to use the request data itself as the encryption key,
such that you can't decrypt the cache unless you actually have the
original user request in hand.

Note: I don't have a Vercel or Node server handy, so this has only been
tested with Wrangler/Cloudflare.