Add tracing, factor out some helper functions, fix a few bugs with Nodejs mode (#68)
This change is a bunch of small tweaks. To summarize:

* Expose a few functions (like app url) out so we can access them when
using it as a library
* Fix a series of streaming/incompat bugs that have accumulated between
cloudflare's workerd runtime and node. For example, Node's `fetch`
decompresses data, while workerd's does not. Also, we weren't properly
closing streams based on how we use Vercel's AI SDK parser, so I just
rewrote that to use our own streaming logic and pulled out the parser
code.
* Add support for tracing.