Disallow issuing temp credential with another temp credential.
We should not allow users to "wrap" a temporary credential with another
temporary credential. This is fine from a security standpoint because
such a credential will fail later while fetching API secrets. However,
allowing this is a confusing and perhaps hard to debug behavior, so we
prefer to fail fast.