Improve type errors (#157)
Certain `fetch` implementations, for example Node/Undici, will throw
`TypeError`s when there's a 500-* error. This is unfortunate because we
lose the mechanics of forwarding these errors along to the client in a
way that they can interpret cleanly.

This change converts those errors into a nicely formatted `Response`, so
at least it's clear what actually happened.