Implement cache-control directives for max-age, no-cache, and no-store (#159)
## Requests
Requests can specify cache behavior and TTL using a limited set of
[Cache-Control](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Cache-Control)
directives. The following request directives are implemented:
* `no-cache`,`no-store` (these are treated as equivalent)
* `max-age` (up to 90 days)

The `x-bt-cache-ttl` header is also added to request a specific TTL
(between 1 second and 90 days)

Note that setting the Cache-Control header will override the behavior of
the `x-bt-use-cache` header.

### Examples:
* Request with `x-bt-cache-ttl: 604800` is equivalent to
`x-bt-use-cache: auto`
* Request with `Cache-control: no-cache` is equivalent to
`x-bt-use-cache: never`
* Request with `Cache-control: no-store` is equivalent to
`x-bt-use-cache: never`
* Request with `Cache-control: max-age=0` is equivalent to
`x-bt-use-cache: never`
* Request with `x-bt-cache-ttl: 77600` will cache requests on supported
paths where temperature is 0 or a seed is set for 90 days
* Request with `x-bt-cache-ttl: 77600`, `Cache-control: max-age=60` will
cache all requests on supported paths for 90 days but will only use
already cached results <= 60 seconds old
* Request with `x-bt-cache-ttl: 77600`, `x-bt-use-cache: always` will
cache all requests on supported paths for 90 days

### Unimplemented request directives
- `max-stale`, `stale-if-error`: Might be useful to implement if the
origin API is down
- `min-fresh`, `no-transform`, `only-if-cached`: Not relevant for a
caching proxy

## Responses
The proxy will respond with basic information about the age of cached
data using the `max-age` directive and the
[Age](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Age)
header. If the request was not placed into the cache, these values will
not be in the response.

### Examples:
* Response with `x-bt-cached: MISS`, `Cache-control: max-age=776000`,
and `Age: 0` missed the cache and was just retrieved from the provider
and cached for 90 days
* Response with `x-bt-cached: MISS` missed the cache and was not cached
* Response with `x-bt-cached: HIT`, `Cache-control: max-age=60`, and
`Age: 15` was retrieved from the cache and will be cached for another 45
seconds

### Unimplemented response directives
- `no-cache`, `no-store`, `must-revalidate`, `proxy-revalidate`,
`must-understand`, `private`, `public`, `stale-while-revalidate`,
`stale-if-error`: Proxy is agnostic to how consumers handle data
  - `s-maxage`: Ignored since proxy doesn't support shared caching
- `immutable`: Maybe respond with this directive if the proxy keeps the
always/auto/never cache semantics
  
- [ ] Followup PR to refactor `timestamp` into non-optional property of
cached data once cache is cycled